bots/photo-to-video-bot/requirements.txt
bubbletea-chat
python-dotenv
opencv-python
pillow
numpy
moviepy
requests
google-generativeai
supabase